Inherit current_author/current_user from host app
When a new entry is created, it is tied to the current_author[:id]. If
no current_author or current_user method is defined in the host
application, the entry cannot be created.

There is still work to be done to improve the UX around this. For
example, it would be nice if the Journal engine would simply *not start*
if no current_author/current_user method is defined in the host
application.