Ensure entry author_ids are never nil
An entry should not be able to be saved if there is no author ID. In the
author_id migration file, I `add_column` and then `change_column` right
after. This is because SQLite produces an error if you add a column with
`null: false` and no `default: <value>`. This simply gets around that
error. *shrug*