Flesh out entries controllers and error messages
This commit ensures that authors can only edit, update, and delete
entries that belong to them. It also ensures that there are useful
flash :notice and :error messages for actions – both on failure and
success.

In the future, the Rails-provided views will be removed or extracted
from this engine. However, for now it's best that they also do not show,
for example, "Edit" buttons in views where the author cannot actually
edit the current @entry.