Add AuthorType (with dynamic field names)
Because Maeku does not provide user authentication, it depends on the
application administrator to add their own. Probably Devise, but maybe
some other thing like Passwordless or Clearance.

Because we don't know what fields will be on the `User`, `Author`, or
whatever the user model is called, we need to account for this and not
make many assumptions in our `AuthorType` design.

We'll do this by looking at the `Maeku.author_class`'s columns and
picking out any non-sensitive ones that are not date/time-oriented.

Right now, I only explicitly look for columns with the term "password"
in them, because all of the sensitive fields provided by Devise include
that word.