Write more robust Rake tasks for all Maeku engines
Previously, the :test Rake task would only test each subproject using
Minitest. But now we have the `maeku_editor` gem, which has both
Minitest and Jest test suites.

This commit creates the all:test, all:minitest, and all:jest tasks so
that we can really, flexibly, test all or some of the suites in this
repository.